finally 0.1

+++ app/Http/Controllers/QL_DatController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers;
 use App\Dat;
+use App\Dat_Web;
 use App\Phuong;
 use App\Quan;
 use App\ThongKeTimKiem;
@@ -13,6 +14,10 @@ class QL_DatController extends Controller {
 	public function getView() {
 		return view('page.quanlydat');
 	}
+	public function getView_DSDatMap()
+    {
+        return view('danhsachdat_map');
+    }
 	public function postThem(Request $request) {
 		$this->validate($request, [
 			'mald' => 'unique:dat,KyHieuLoDat',
@@ -21,42 +26,52 @@ class QL_DatController extends Controller {
 		]);
 		$dat = new Dat;
 		//
+		if($dat->kiemtraMap($request->map)) {
 		// xử lý ảnh
-		$file = $request->file('image');
-		$i = 0;
-		$tam = '';
-		foreach ($file as $image) {
-			if ($i <= 2) {
-				$input['imagename'] = $request->mald . '_' . ++$i . '.' . $image->getClientOriginalExtension();
-				$destinationPath = public_path('/img');
-				$image->move($destinationPath, $input['imagename']);
-				$tam = $input['imagename'] . ';' . $tam;
+			$file = $request->file('image');
+			$i = 0;
+			$tam = '';
+			foreach ($file as $image) {
+				if ($i <= 2) {
+					$input['imagename'] = $request->mald . '_' . ++$i . '.' . $image->getClientOriginalExtension();
+					$destinationPath = public_path('/img');
+					$image->move($destinationPath, $input['imagename']);
+					$tam = $input['imagename'] . ';' . $tam;
+				}
 			}
-		}
 		//
-		$dat->HinhAnh = $tam;
-		$dt = ($request->dai) * ($request->rong) + (0.5 * ($request->nohau) * ($request->dai));
-		$gia = $dt * ($request->dongia);
-		$dat->KyHieuLoDat = $request->mald;
-		$dat->SoHuu = $request->sohuu;
-		$dat->TrangThai = '0';
-		$dat->LuotXem = '0';
-		$dat->DiaChi = $request->diachi;
-		$dat->Quan = $request->quan;
-		//$dat->ThanhPho = $request->tp;
-		$dat->Gia = $gia;
-		$dat->Dai = $request->dai;
-		$dat->Rong = $request->rong;
-		$dat->DonGia = $request->dongia;
-		$dat->NoHau = $request->nohau;
-		$dat->DienTich = $dt;
-		$dat->Huong = $request->huong;
-		$dat->GhiChu = $request->ghichu;
-		$dat->save();
-		return redirect('page/quanlydat')->with('thongbao', 'Thêm thành công thông tin lô đất !');
+			$dat->HinhAnh = $tam;
+			$dt = ($request->dai) * ($request->rong) + (0.5 * ($request->nohau) * ($request->dai));
+			$gia = $dt * ($request->dongia);
+			$dat->KyHieuLoDat = $request->mald;
+			$dat->SoHuu = $request->sohuu;
+			$dat->TrangThai = $request->trangthai;
+			$dat->LuotXem = '0';
+			$dat->Map = $request->map;
+			$dat->DiaChi = $request->diachi;
+			$dat->Phuong = $request->phuong;
+			//$dat->ThanhPho = $request->tp;
+			$dat->Gia = $gia;
+			$dat->Dai = $request->dai;
+			$dat->Rong = $request->rong;
+			$dat->DonGia = $request->dongia;
+			$dat->NoHau = $request->nohau;
+			$dat->DienTich = $dt;
+			$dat->Huong = $request->huong;
+			$dat->GhiChu = $request->ghichu;
+			$dat->save();
+			return redirect('page/quanlydat')->with('thongbao', 'Thêm thành công thông tin lô đất !');
+		}
+		return redirect('page/quanlydat')->with('canhbao', 'Địa chỉ đã có trong hệ thống');
 	}
 	public function getXoa($id) {
 		$dat = Dat::find($id);
+		if ($dat->TrangThai == 1) {
+			return redirect('page/quanlydat')->with('canhbao', 'Lô đất đang đợi giao dịch, không thể xóa !');
+		}
+		if ($dat->TrangThai == 2) {
+			return redirect('page/quanlydat')->with('canhbao', 'Lô đất đã có giao dịch trước đó, không thể xóa !');
+		}
 		$s = $dat->HinhAnh;
 		$m = explode(';', $s);
 		for ($a = 0; $a < (count($m) - 1); $a++) {
@@ -77,52 +92,58 @@ class QL_DatController extends Controller {
 			'dongia.min' => 'Đơn giá lô đất phải lớn hơn 0',
 			'nohau.min' => 'Độ nở hậu lô đất phải lớn hơn hoặc bằng 0',
 		]);
-		$dat = Dat::find($request->iddat);
-		$file = $request->file('image');
-		$i = 0;
-		$tam = '';
-		if ($file != NULL) {
-			$s = $dat->HinhAnh;
-			$m = explode(';', $s);
-			for ($a = 0; $a < (count($m) - 1); $a++) {
-				Storage::delete($m[$a]);
-			}
-			foreach ($file as $image) {
-				if ($i <= 2) {
-					$input['imagename'] = $request->mald . '_' . ++$i . '.' . $image->getClientOriginalExtension();
-					$destinationPath = public_path('/img');
-					$image->move($destinationPath, $input['imagename']);
-					$tam = $input['imagename'] . ';' . $tam;
+		$check = new Dat();
+		if($check->kiemtraMap($request->Map)) {
+			$dat = Dat::find($request->iddat);
+			$file = $request->file('image');
+			$i = 0;
+			$tam = '';
+			if ($file != NULL) {
+				$s = $dat->HinhAnh;
+				$m = explode(';', $s);
+				for ($a = 0; $a < (count($m) - 1); $a++) {
+					Storage::delete($m[$a]);
 				}
+				foreach ($file as $image) {
+					if ($i <= 2) {
+						$input['imagename'] = $request->mald . '_' . ++$i . '.' . $image->getClientOriginalExtension();
+						$destinationPath = public_path('/img');
+						$image->move($destinationPath, $input['imagename']);
+						$tam = $input['imagename'] . ';' . $tam;
+					}
+				}
+				$dat->HinhAnh = $tam;
 			}
-			$dat->HinhAnh = $tam;
-		}
-		$dt = ($request->dai) * ($request->rong) + (0.5 * ($request->nohau) * ($request->dai));
-		$gia = $dt * ($request->dongia);
-		$dat->KyHieuLoDat = $request->mald;
-		$dat->SoHuu = $request->sohuu;
-		$dat->DiaChi = $request->diachi;
-		$dat->Quan = $request->quan;
-		//$dat->ThanhPho = $request->tp;
-		$dat->Gia = $gia;
-		$dat->Dai = $request->dai;
-		$dat->Rong = $request->rong;
-		$dat->DonGia = $request->dongia;
-		$dat->NoHau = $request->nohau;
-		$dat->DienTich = $dt;
-		$dat->Huong = $request->huong;
-		$dat->GhiChu = $request->ghichu;
-		$dat->save();
+			$dt = ($request->dai) * ($request->rong) + (0.5 * ($request->nohau) * ($request->dai));
+			$gia = $dt * ($request->dongia);
+			$dat->KyHieuLoDat = $request->mald;
+			$dat->SoHuu = $request->sohuu;
+			$dat->DiaChi = $request->diachi;
+			$dat->Phuong = $request->phuong;
+			//$dat->ThanhPho = $request->tp;
+			$dat->Gia = $gia;
+			$dat->Dai = $request->dai;
+			$dat->Map = $request->map;
+			$dat->Rong = $request->rong;
+			$dat->DonGia = $request->dongia;
+			$dat->NoHau = $request->nohau;
+			$dat->DienTich = $dt;
+			$dat->Huong = $request->huong;
+			$dat->GhiChu = $request->ghichu;
+			//$dat->save();
 
-		return redirect('page/quanlydat')->with('thongbao', 'Cập nhật thành công thông tin lô đất !');
+			return redirect('page/quanlydat')->with('thongbao', 'Cập nhật thành công thông tin lô đất !');
+		}
+		return redirect('page/quanlydat')->with('canhbao', 'Địa chỉ đã có trong hệ thống');
 	}
 	public function getTim(Request $r) {
 		$dat = new Dat;
 		echo $dat->timdat($r->name);
 	}
 	public function getView_DSDat() {
-		$datFromOtherPage = $this->findDatFromOtherWeb();
-		return view('danhsachdat', ['resultFromWeb'=>$datFromOtherPage]);
+		$dat = new Dat;
+		return view('danhsachdat');
+		$trangthai = $dat->getTrangThaiChucNang();		
 	}
 	public function timDat_ban(Request $r) {
 		$dt = $r->dt;
@@ -132,8 +153,8 @@ class QL_DatController extends Controller {
 		$tp = 1;
 		$quan = $r->quan;
 		//search dat
-		$a = new Dat();
-		$kq = $a->timdat_ban($quan, $tp, $gia, $dt, $huong);
+		$dat = new Dat();
+		$kq = $dat->timdat_ban($quan, $tp, $gia, $dt, $huong);
 		// update db thongketimkiem
 		if ($quan != 0 || $huong != "A") {
 			$thongketimkiem = new ThongKeTimKiem();
@@ -143,10 +164,13 @@ class QL_DatController extends Controller {
 			$thongketimkiem->deleteAfterOneYear();
 		}
 		// tim dat tu trang web khac
-		$datFromOtherPage = $this->findDatFromOtherWeb($quan, $gia, $dt);
-		session(['datFromOtherPage' => $datFromOtherPage]);
-		//return view
-		return view('danhsachdat_kqtim', ['kq'=>$kq, 'resultFromWeb'=>$datFromOtherPage]);
+		$trangthai = $dat->getTrangThaiChucNang();
+		if( $trangthai == 0 && count($kq) < 9 ) {
+			$datFromOtherPage = $this->findDatFromOtherWeb($quan, $gia, $dt);
+			return view('danhsachdat_kqtim', ['kq'=>$kq, 'resultFromWeb'=>$datFromOtherPage]);
+		} else {
+			return view('danhsachdat_kqtim', ['kq'=>$kq]);
+		}
 	}
 	public function timQuan(Request $r) {
 		$quan = new Quan;
@@ -208,12 +232,13 @@ class QL_DatController extends Controller {
         }
 		$html = file_get_html('http://www.muabannhadat.vn/dat-ban-3515/tp-ho-chi-minh-'.$tagQuan.'?aral='.$gia.$dt , FILE_USE_INCLUDE_PATH);
 		$array = array();
-		$count = 5;
+		$count = 6;
 		for($i=0; $i<$count; $i++) {
 			$links = array();
 			foreach($html->find('//*[@id="MainContent_ctlList_ctlResults_repList_ctl00_'.$i.'_divListingInformationTitle_'.$i.'"]/a') as $a) {
 				$html_sub = file_get_html('http://www.muabannhadat.vn'.$a->href, FILE_USE_INCLUDE_PATH);
 				$linkGet = 'http://www.muabannhadat.vn'.$a->href;
+				//
 				$links[] = str_replace('/dat-ban-dat-tho-cu-3532/','',$a->href);
 				// get tieu de
 				foreach($html_sub->find('//*[@id="ctl01"]/div[5]/div[2]/div/div/div/div[1]') as $a) {
@@ -224,7 +249,7 @@ class QL_DatController extends Controller {
 				foreach($html_sub->find('//*[@id="ctl01"]/div[5]/div[3]/div/div/div[4]/div[1]/div[1]/div[1]') as $a) {
 					$vitri = explode('|', $a->plaintext);
 					$links[] = $vitri[2];
-					$links[] = $vitri[1].', '.$vitri[2].', '.$vitri[2];
+					$links[] = $vitri[1].', '.$vitri[2];
 					break;
 				}
 				// get vi tri map
@@ -244,7 +269,11 @@ class QL_DatController extends Controller {
 				}
 				// get mo ta
 				foreach($html_sub->find('//*[@id="Description"]') as $a) {
-					$links[] = $a->plaintext;
+					if(strlen($a->plaintext) > 1000) {
+						$links[] = $this->substr($a->plaintext, 1000);
+					} else {
+						$links[] = $a->plaintext;
+					}
 					break;
 				}
 				// get hinh
@@ -256,123 +285,55 @@ class QL_DatController extends Controller {
 					$links[] = $a->plaintext;
 					break;
 				}
-				// get huong 
 				$links[] = $linkGet;
+
+				// save to DB
 				if (count($links) >= 11) {
-					$array[$i] = $links;
-				}
-				else {
+					$dat = new Dat_Web();
+					if($dat->checkLink($links[0])) {
+						$dat->link = $links[0];
+						$dat->TrangThai = 0;
+						$dat->Gia = $links[6];
+						$dat->DiaChi = $links[3];
+						$dat->DienTich = $links[5];
+						$dat->MoTa = $links[7];
+						$dat->Map = $links[4];
+						$dat->Hinh = $links[8];
+						$dat->save();
+						$array[$i] = $links;
+					} else {
+						$id = $dat->findIdByLink($links[0]);
+						$dat_web = Dat_Web::find($id);
+						if ($dat_web->TrangThai == 2 || $dat_web->TrangThai == 1) {
+							$count++;
+						} else {
+							$array[$i] = $links;
+						}
+					}
+				} else {
 					$count++;
 				}
-				// get hinh 
 			}
 		}
 		return $array;
 	}
-
-	public function getDatFromNhaDatNet($quan, $gia, $dt) {
-		if ($quan != 0) {
-			$timQuan = Quan::find($quan);
-			$tagQuan = $timQuan->Tag_nhadat;
-		}
-		else {
-			$tagQuan = 's59';
-		}
-		switch ($dt) {
-            case '1':
-                $dt = 'dien-tich:0-50';
-                break;
-            case '2':
-                $dt = 'dien-tich:50-100';
-                break;
-            case '3':
-                $dt = 'dien-tich:100-150';
-                break;
-            case '4':
-                $dt = 'dien-tich:150-200';
-                break;
-            default:
-                $dt = '';
-                break;
-        }
-        switch ($gia) {
-            case '1':
-                $gia = 'gia:0-800';
-                break;
-            case '2':
-                $gia = 'gia:800-1500';
-                break;
-            case '3':
-                $gia = 'gia:1500-2500';
-                break;
-            case '4':
-                $gia = 'gia:2500-4000';
-                break;
-            default:
-                $gia = '';
-                break;
-        }
-		$html = file_get_html('http://www.muabannhadat.vn/dat-ban-3515/tp-ho-chi-minh-'.$tagQuan.'?aral='.$gia.$dt , FILE_USE_INCLUDE_PATH);
-		$array = array();
-		$count = 5;
-		for($i=0; $i<$count; $i++) {
-			$links = array();
-			foreach($html->find('//*[@id="MainContent_ctlList_ctlResults_repList_ctl00_'.$i.'_divListingInformationTitle_'.$i.'"]/a') as $a) {
-				$html_sub = file_get_html('http://www.muabannhadat.vn'.$a->href, FILE_USE_INCLUDE_PATH);
-				$linkGet = 'http://www.muabannhadat.vn'.$a->href;
-				$links[] = str_replace('/dat-ban-dat-tho-cu-3532/','',$a->href);
-				// get tieu de
-				foreach($html_sub->find('//*[@id="ctl01"]/div[5]/div[2]/div/div/div/div[1]') as $a) {
-					$links[] = $a->plaintext;
-					break;
-				}
-				// get dia chi
-				foreach($html_sub->find('//*[@id="ctl01"]/div[5]/div[3]/div/div/div[4]/div[1]/div[1]/div[1]') as $a) {
-					$vitri = explode('|', $a->plaintext);
-					$links[] = $vitri[2];
-					$links[] = $vitri[1].', '.$vitri[2].', '.$vitri[2];
-					break;
-				}
-				// get vi tri map
-				foreach($html_sub->find('//*[@id="MainContent_ctlDetailBox_lblMapLink"]/a') as $a) {
-					$links[] = str_replace('https://maps.google.com/?q=loc:', '', $a->href);
-					break;
-				}
-				// get dien tich
-				foreach($html_sub->find('//*[@id="ctl01"]/div[5]/div[3]/div/div/div[4]/div[1]/div[2]/div/div[2]/div[1]/table/tbody/tr[2]/td') as $a) {
-					$links[] = $a->plaintext;
-					break;
-				}
-				// get gia tien
-				foreach($html_sub->find('//*[@id="MainContent_ctlDetailBox_lblPrice"]') as $a) {
-					$links[] = $a->plaintext;
-					break;
-				}
-				// get mo ta
-				foreach($html_sub->find('//*[@id="Description"]') as $a) {
-					$links[] = $a->plaintext;
-					break;
-				}
-				// get hinh
-				foreach($html_sub->find('.swipebox') as $a) {
-					$links[] = $a->href;
-					break;
-				}
-				foreach($html_sub->find('//*[@id="MainContent_ctlDetailBox_lblFengShuiDirection"]') as $a) {
-					$links[] = $a->plaintext;
-					break;
-				}
-				// get huong 
-				$links[] = $linkGet;
-				if (count($links) >= 11) {
-					$array[$i] = $links;
-				}
-				else {
-					$count++;
-				}
-				// get hinh 
-			}
-		}
-		return $array;
+	function substr($str, $length, $minword = 3)
+	{
+	    $sub = '';
+	    $len = 0;
+	    foreach (explode(' ', $str) as $word) {
+	        $part = (($sub != '') ? ' ' : '') . $word;
+	        $sub .= $part;
+	        $len += strlen($part);
+	        if (strlen($word) > $minword && strlen($sub) >= $length) {
+	            break;
+	        }
+	    }
+	    return $sub . (($len < strlen($str)) ? '...' : '');
+	}
+	public function batTatChucNang (Request $r) {
+		$dat = new Dat;
+		$dat->batTatChucNang($r->check);
+		return $r->check;
 	}
 }
